chức năng bình luận

+++ database/migrations/2025_04_12_152414_create_binh_luan_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('binh_luan', function (Blueprint $table) {
             $table->id('ID_BinhLuan');
-            $table->tinyInteger('DiemDanhGia');
+            $table->float('DiemDanhGia', 3, 1);
             $table->unsignedBigInteger('ID_TaiKhoan');
             $table->unsignedBigInteger('ID_Phim');
             $table->foreign('ID_TaiKhoan')->references('ID_TaiKhoan')->on('tai_khoan');
